Assets file moving (without conflict resolution)

+++ src/services/AssetTransforms.php
@@ -436,6 +436,13 @@ class AssetTransforms extends Component
 			/** @var array $matchFound */
 			$from = $file->getFolder()->path.$this->getTransformSubpath($file, AssetTransformIndex::create($matchFound));
 			$to = $file->getFolder()->path.$this->getTransformSubpath($file, $index);
+
+			// Sanity check
+			if ($volume->fileExists($to))
+			{
+				return;
+			}
+
 			$volume->copyFile($from, $to);
 		}
 		else
@@ -619,7 +626,7 @@ class AssetTransforms extends Component
 	 */
 	public function deleteTransformIndexDataByFileId($fileId)
 	{
-		Craft::$app->getDb()->createCommand()->delete('{{%assettransformindex}}', 'fileId = :fileId', [':fileId' => $fileId]);
+		Craft::$app->getDb()->createCommand()->delete('{{%assettransformindex}}', 'fileId = :fileId', [':fileId' => $fileId])->execute();
 	}
 
 	/**
@@ -631,7 +638,7 @@ class AssetTransforms extends Component
 	 */
 	public function deleteTransformIndex($indexId)
 	{
-		Craft::$app->getDb()->createCommand()->delete('{{%assettransformindex}}', 'id = :id', [':id' => $indexId]);
+		Craft::$app->getDb()->createCommand()->delete('{{%assettransformindex}}', 'id = :id', [':id' => $indexId])->execute();
 	}
 	/**
 	 * Get a thumb server path by file model and size.
@@ -922,18 +929,17 @@ class AssetTransforms extends Component
 	/**
 	 * Delete created transforms for a file.
 	 *
-	 * @param Asset $file
+	 * @param Asset $asset
 	 */
-	public function deleteCreatedTransformsForFile(Asset $file)
+	public function deleteCreatedTransformsForFile(Asset $asset)
 	{
-		$indexModels = $this->getAllCreatedTransformsForFile($file);
+		$transformIndexes = $this->getAllCreatedTransformsForFile($asset);
 
-		$volume = $file->getVolume();
+		$volume = $asset->getVolume();
 
-		foreach ($indexModels as $index)
+		foreach ($transformIndexes as $transformIndex)
 		{
-			// TODO Transform handling!
-			//$volume->deleteTransform($file, $index);
+			$volume->deleteFile($asset->getFolder()->path.Craft::$app->assetTransforms->getTransformSubpath($asset, $transformIndex));
 		}
 	}
 
@@ -1058,6 +1064,14 @@ class AssetTransforms extends Component
 		}
 
 		$volume = $file->getVolume();
+		$transformPath = $file->getFolder()->path . $this->getTransformSubpath($file, $index);
+
+		// Already created. Relax, grasshopper!
+		if ($volume->fileExists($transformPath))
+		{
+			return;
+		}
+
 		$imageSource = $file->getTransformSource();
 		$quality = $transform->quality ? $transform->quality : Craft::$app->config->get('defaultImageQuality');
 
@@ -1089,7 +1103,6 @@ class AssetTransforms extends Component
 		$image->saveAs($createdTransform);
 
 		clearstatcache(true, $createdTransform);
-		$transformPath = $file->getFolder()->path . $this->getTransformSubpath($file, $index);
 
 		$stream = fopen($createdTransform, "r");
 		$volume->createFileByStream($transformPath, $stream);
